+++ neo/d3xp/gamesys/SaveGame.cpp
@@ -32,8 +32,6 @@ If you have questions concerning this license or the applicable additional terms
 
 #include "../Game_local.h"
 
-#include "TypeInfo.h"
-
 /*
 Save game related helper classes.
 
@@ -291,7 +289,7 @@ void idSaveGame::WriteString( const char *string ) {
 	stringHash.Add( hash, stringTable.Num() - 1 );
 
 	WriteInt( curStringTableOffset );
-	curStringTableOffset += ( strlen( string ) + 4 );
+	curStringTableOffset += (int) ( strlen( string ) + 4 );
 }
 
 /*
