+++ neo/d3xp/script/Script_Thread.cpp
@@ -623,7 +623,7 @@ void idThread::KillThread( const char *name ) {
 	if ( ptr ) {
 		len = ptr - name;
 	} else {
-		len = strlen( name );
+		len = (int)strlen( name );
 	}
 
 	// kill only those threads whose name matches name
@@ -1693,7 +1693,7 @@ idThread::Event_StrLen
 void idThread::Event_StrLen( const char *string ) {
 	int len;
 
-	len = strlen( string );
+	len = (int)strlen( string );
 	idThread::ReturnInt( len );
 }
 
@@ -1703,7 +1703,7 @@ idThread::Event_StrLeft
 ================
 */
 void idThread::Event_StrLeft( const char *string, int num ) {
-	int len;
+	size_t len;
 
 	if ( num < 0 ) {
 		idThread::ReturnString( "" );
@@ -1726,7 +1726,7 @@ idThread::Event_StrRight
 ================
 */
 void idThread::Event_StrRight( const char *string, int num ) {
-	int len;
+	size_t len;
 
 	if ( num < 0 ) {
 		idThread::ReturnString( "" );
@@ -1748,7 +1748,7 @@ idThread::Event_StrSkip
 ================
 */
 void idThread::Event_StrSkip( const char *string, int num ) {
-	int len;
+	size_t len;
 
 	if ( num < 0 ) {
 		idThread::ReturnString( string );
@@ -1780,7 +1780,7 @@ void idThread::Event_StrMid( const char *string, int start, int num ) {
 	if ( start < 0 ) {
 		start = 0;
 	}
-	len = strlen( string );
+	len = (int)strlen( string );
 	if ( start > len ) {
 		start = len;
 	}
