+++ neo/idlib/Heap.cpp
@@ -42,11 +42,11 @@ If you have questions concerning this license or the applicable additional terms
 Mem_Alloc16
 ==================
 */
-void * Mem_Alloc16( const int size, const memTag_t tag ) {
+void * Mem_Alloc16( const size_t size, const memTag_t tag ) {
 	if ( !size ) {
 		return NULL;
 	}
-	const int paddedSize = ( size + 15 ) & ~15;
+	const size_t paddedSize = ( size + 15 ) & ~15;
 	return _aligned_malloc( paddedSize, 16 );
 }
 
@@ -67,9 +67,9 @@ void Mem_Free16( void *ptr ) {
 Mem_ClearedAlloc
 ==================
 */
-void * Mem_ClearedAlloc( const int size, const memTag_t tag ) {
+void * Mem_ClearedAlloc( const size_t size, const memTag_t tag ) {
 	void * mem = Mem_Alloc( size, tag );
-	SIMDProcessor->Memset( mem, 0, size );
+	SIMDProcessor->Memset( mem, 0, (int)size );
 	return mem;
 }
