+++ neo/idlib/Heap.h
@@ -49,34 +49,34 @@ static const int MAX_TAGS = 256;
 
 
 
-void *		Mem_Alloc16( const int size, const memTag_t tag );
+void *		Mem_Alloc16( const size_t size, const memTag_t tag );
 void		Mem_Free16( void *ptr );
 
-ID_INLINE void *	Mem_Alloc( const int size, const memTag_t tag ) { return Mem_Alloc16( size, tag ); }
+ID_INLINE void *	Mem_Alloc( const size_t size, const memTag_t tag ) { return Mem_Alloc16( size, tag ); }
 ID_INLINE void		Mem_Free( void *ptr ) { Mem_Free16( ptr ); }
 
-void *		Mem_ClearedAlloc( const int size, const memTag_t tag );
+void *		Mem_ClearedAlloc( const size_t size, const memTag_t tag );
 char *		Mem_CopyString( const char *in );
 
-ID_INLINE void *operator new( size_t s ) {
+ID_INLINE void *operator new(size_t s ) {
 	return Mem_Alloc( s, TAG_NEW );
 }
 ID_INLINE void operator delete( void *p ) {
 	Mem_Free( p );
 }
-ID_INLINE void *operator new[]( size_t s ) {
+ID_INLINE void *operator new[](size_t s ) {
 	return Mem_Alloc( s, TAG_NEW );
 }
 ID_INLINE void operator delete[]( void *p ) {
 	Mem_Free( p );
 }
-ID_INLINE void *operator new( size_t s, memTag_t tag ) {
+ID_INLINE void *operator new(size_t s, memTag_t tag ) {
 	return Mem_Alloc( s, tag );
 }
 ID_INLINE void operator delete( void *p, memTag_t tag ) {
 	Mem_Free( p );
 }
-ID_INLINE void *operator new[]( size_t s, memTag_t tag ) {
+ID_INLINE void *operator new[](size_t s, memTag_t tag ) {
 	return Mem_Alloc( s, tag );
 }
 ID_INLINE void operator delete[]( void *p, memTag_t tag ) {
