+++ neo/ui/Winvar.cpp
@@ -53,9 +53,9 @@ void idWinVar::SetGuiInfo(idDict *gd, const char *_name) {
 void idWinVar::Init(const char *_name, idWindow *win) {
 	idStr key = _name;
 	guiDict = NULL;
-	int len = key.Length();
+	size_t len = key.Length();
 	if (len > 5 && key[0] == 'g' && key[1] == 'u' && key[2] == 'i' && key[3] == ':') {
-		key = key.Right(len - VAR_GUIPREFIX_LEN);
+		key = key.Right((int)(len - VAR_GUIPREFIX_LEN));
 		SetGuiInfo( win->GetGui()->GetStateDict(), key );
 		win->AddUpdateVar(this);
 	} else {
